cosmetics
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5078 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/SubPic/CoordGeom.cpp
@@ -366,7 +366,7 @@ float Ray::GetDistanceFrom(Ray& r)
 {
 	float t = (d | r.d);
 	if (IsZero(t)) {
-		return (float)-BIGNUMBER;    // plane is paralell to the ray, return -infinite
+		return -BIGNUMBER;    // plane is paralell to the ray, return -infinite
 	}
 	return (((r.p - p) | r.d) / t);
 }
